+++ build_test.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+# Set up test environment
+python3 -m pip install --upgrade wheel setuptools pip
+python3 setup.py install
+python3 -m pip install -U pytest
+
+# Add flag for test environment
+export IS_CICD=true
+
+# Force install of test dependencies before running pytest (should happen
+#   automatically but sometimes fails)
+python3 -m pip install .[test]
+python3 -m pytest