+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file. Please do f
 _“Yesterday I was clever, so I wanted to change the world. Today I am wise, so I am changing myself.”_ - Rumi
 
 
+## [0.1.10] - 2021-06-08
+### Testing Updates
+- Validation added to check http & https URLs in notebooks and markdown documents.
+- Windows added as agent to CI pipeline (to test on multiple OS)
+
 ## [0.1.9] - 2021-04-19
 ### Reporting Updates
 - Both standard and stratified reporting functions are now contained in a single module, and new APIs have been put in place to facilitate both classification and regression reports